# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Teranga GESCRIM API",
        "description": "API REST pour le système de Gestion de la Criminalité et de la Délinquance — Direction de la Sécurité Publique (DSP) du Sénégal.",
        "contact": {
            "name": "NAMASTECH",
            "email": "admin@gescrim.sn"
        },
        "license": {
            "name": "Propriétaire",
            "url": "https://gescrim.sn"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/api/sync/batch": {
            "post": {
                "tags": [
                    "Synchronisation"
                ],
                "summary": "Synchroniser les données offline par lot",
                "description": "Envoie en une seule requête toutes les infractions et accidents saisis en mode hors-ligne.",
                "operationId": "17b478c48719a9159824836a063841b9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "infractions": {
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/Infraction"
                                        }
                                    },
                                    "accidents": {
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/Accident"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Synchronisation réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "infractions_synced": {
                                            "type": "integer"
                                        },
                                        "accidents_synced": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/sync/status": {
            "get": {
                "tags": [
                    "Synchronisation"
                ],
                "summary": "État de synchronisation",
                "description": "Retourne le nombre d'éléments en attente et synchonisés.",
                "operationId": "7bd5ab03d3512ac159a2922b6e67f5af",
                "responses": {
                    "200": {
                        "description": "Statut de sync"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Erreur rencontrée"
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Infraction": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type_infraction_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "service_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "annee": {
                        "type": "integer",
                        "example": 2026
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-05-17"
                    },
                    "lieu": {
                        "type": "string",
                        "example": "Sandaga, Dakar"
                    },
                    "commune_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "issue": {
                        "type": "string",
                        "example": "Constatée"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 14.6928
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": -17.4467
                    },
                    "description": {
                        "type": "string",
                        "example": "Usage de téléphone au volant constaté lors d'un contrôle."
                    },
                    "sync_status": {
                        "type": "string",
                        "example": "synced"
                    }
                },
                "type": "object"
            },
            "Accident": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "example": "corporel"
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-05-17"
                    },
                    "lieu": {
                        "type": "string",
                        "example": "Vdn, Dakar"
                    },
                    "commune_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "service_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "moyen": {
                        "type": "string",
                        "example": "Scooter vs Voiture"
                    },
                    "cause_probable": {
                        "type": "string",
                        "example": "Non respect de la priorité"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 14.7228
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": -17.4567
                    },
                    "description": {
                        "type": "string",
                        "example": "Accident corporel léger."
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Entrez votre token JWT (sans le mot Bearer)"
            }
        }
    },
...
}